day_55/requirements.txt
google-generativeai==0.8.5
python-dotenv==0.19.0